+++ aioxmpp/__init__.py
@@ -100,6 +100,7 @@ from .pubsub import PubSubClient  # NOQA
 from .shim import SHIMService  # NOQA
 from .adhoc import AdHocClient, AdHocServer  # NOQA
 from .avatar import AvatarClient, AvatarServer  # NOQA
+from .carbons import CarbonsClient  # NOQA
 
 
 def set_strict_mode():